app/api/requirements.txt
flask
PyJWT
requests
google-cloud-storage
google-cloud-bigquery
google-cloud-logging
flasgger
waitress